Adding README, UNLICENSE, and a C++ program.
Completed README.md, added the dynamic_list.cpp program, removed the
dirc.c program (since it was too short and it was unknown if it was
compatible with the license of this repository). Added the LICENSE
(UNLICENSE). Added the CONTRIBUTING.md file (but it is blank), and also
added the list.md file (which is also blank).